requirements.txt
fastapi
uvicorn
langchain
langgraph
nacos-sdk-python
httpx
sseclient-py
mcp[sse]
psycopg[binary]
sqlalchemy
langgraph-checkpoint-postgres
psycopg-pool
python-dotenv